+++ Core/Native/User32.cs
@@ -141,11 +141,6 @@ internal static partial class User32
         IntPtr hWnd, uint Msg, IntPtr wParam, IntPtr lParam,
         uint fuFlags, uint uTimeout, out IntPtr lpdwResult);
 
-    [LibraryImport("user32.dll", EntryPoint = "SendMessageTimeoutW", SetLastError = true)]
-    internal static partial IntPtr SendMessageTimeoutTitleBarInfo(
-        IntPtr hWnd, uint Msg, IntPtr wParam, ref TITLEBARINFOEX lParam,
-        uint fuFlags, uint uTimeout, out IntPtr lpdwResult);
-
     [LibraryImport("user32.dll")]
     internal static partial IntPtr DefWindowProcW(IntPtr hWnd, uint Msg, IntPtr wParam, IntPtr lParam);